rd-226 implement switcher attribute for code
